mcp-crowdstrike-server/pyproject.toml
[project]
name = "mcp-crowdstrike"
version = "0.1.0"
description = "A Model Context Protocol server that connects to CrowdStrike's Falcon API for usage by LLMs"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "CrowdStrike" }]
maintainers = [{ name = "Carlos Matos", email = "carlos.matos@crowdstrike.com" }]
keywords = ["mcp", "llm", "security", "intel"]
license = { text = "MIT" }
dependencies = [
    "crowdstrike-falconpy>=1.5.0",
    "mcp>=1.7.1",
    "python-dotenv>=1.1.0",
]
[project.optional-dependencies]
test = [
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
mcp-crowdstrike = "mcp_crowdstrike.server:main"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
